Turns --print into a boolean option.
If you try something like:

    bookmaker --print input.pdf

it expects that input.pdf is the printer and that there is no input file.
To make it work, one has to

    bookmaker input.pdf --print

which is non-obvious and seems to be in opposition with the automatically
printed help.

When reading the code, I had expected that required arguments (like
input) would be filled, while options arguments, like that to print,
would be filled only when the required arguments would also be filled.

Until a solution is found, I have turned --print into a boolean option
(defaults to False), which when True (by specifying --print) will print
to the default printer. However, this does not allow a printer to be
specified.

See pull request 5.